fix: get node name link working
Changed the node-name-as-link functionality to be implemented as a click handler on the name's text element, rather than an SVG <a> tag, as I could not get the latter to work properly (i.e. the navigation would often not occur), even after much experimentation.
Used zoom-in and -out mouse cursors for the collapsible nodes in the tree.